parking_system/requirements.txt
Django==3.2.4
gunicorn==20.1.0
psycopg2-binary==2.9.1
numpy==1.23
ultralytics==8.2.32
matplotlib==3.5.1
python-dotenv==0.19.2
easyocr
liqpay-sdk-python3
telebot